zram: remove zram_sysfs file (v2)
Move zram sysfs code to zram drv and remove zram_sysfs.c
file. This gives ability to make static a number of previously
exported zram functions, used from zram sysfs, e.g. internal zram
zram_meta_alloc/free(). We also can drop zram_drv wrapper
functions, used from zram sysfs:
e.g. zram_reset_device()/__zram_reset_device() pair.

v2: as suggested by Greg K-H, move MODULE description to the
bottom of the file.

Git-commit: 9b3bb7abcdf2df0f1b2657e6cbc9d06bc2b3b36f
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
